+++ .storybook/preview.js
@@ -5,7 +5,16 @@ import React from 'react'
 import { DevGridDecorator } from './decorators/dev-grid'
 import { DocsPage, DocsContainer } from '@storybook/addon-docs'
 import { themes } from '@storybook/theming';
-import { FileUrl } from 'storybook/file-url'
+
+const tokenContext = require.context(
+  '!!raw-loader!../assets/stylesheets',
+  true,
+  /.\.(css|less|scss|svg)$/
+);
+
+const tokenFiles = tokenContext.keys().map(function (filename) {
+  return { filename: filename, content: tokenContext(filename).default };
+});
 
 export const parameters = {
   docs: {
@@ -17,7 +26,7 @@ export const parameters = {
     storySort: {
       order: [
         'Documentation',
-        ['Introduction', 'Usage', 'Design'],
+        ['Introduction', 'Usage', 'Design', 'Design Token'],
         'Helpers',
         'Accessibility',
         'Graphics',
@@ -34,7 +43,11 @@ export const parameters = {
   controls: {
     expanded: true,
     sort: 'requiredFirst',
- },
+  },
+  designToken: {
+    defaultTab: 'Colors',
+    files: tokenFiles
+  },
 }
 
 export const decorators = [