+++ assets/javascripts/kitten/components/form/text-input/index.js
@@ -9,14 +9,14 @@ import classNames from 'classnames'
 
 const StyledInput = styled.input`
   --input-padding-horizontal: ${pxToRem(15)};
-  --input-border-width: ${pxToRem(2)};
+  --input-border-width: var(--border-width);
   --input-height: ${pxToRem(50)};
 
   font-size: ${stepToRem(-1)};
   line-height: 1em;
   ${TYPOGRAPHY.fontStyles.light};
   box-sizing: border-box;
-  border-width: ${pxToRem(2)}; /* IE11 */
+  border-width: ${pxToRem(1)}; /* IE11 */
   border-width: var(--input-border-width);
   border-style: solid;
   border-radius: 0;
@@ -24,7 +24,7 @@ const StyledInput = styled.input`
   appearance: none;
   background-color: ${COLORS.background1};
   color: ${COLORS.font1};
-  border-color: ${COLORS.line1};
+  border-color: var(--color-grey-400);
   height: var(--input-height);
   min-height: var(--input-height);
 
@@ -138,7 +138,7 @@ const StyledInput = styled.input`
   }
 
   &:hover {
-    border-color: ${COLORS.line2};
+    border-color: var(--color-grey-500);
   }
 
   &:focus {