+++ assets/javascripts/kitten/components/organisms/tables/double-entry-table/styles.js
@@ -60,7 +60,7 @@ export const Styles = styled.div`
   .DoubleEntryTable__Column {
     height: ${pxToRem(68)};
     padding: 0 ${pxToRem(20)};
-    border: ${pxToRem(2)} solid rgba(0, 0, 0, 0.067);
+    border: var(--border-width) solid var(--color-grey-400);
     font-size: ${stepToRem(-1)};
     line-height: 1.3;
     box-sizing: border-box;
@@ -70,7 +70,7 @@ export const Styles = styled.div`
 
     @media screen and (min-width: 0\0) {
       /* IE */
-      border: ${pxToRem(2)} solid ${COLORS.line1};
+      border: ${pxToRem(1)} solid ${COLORS.line2};
     }
 
     &.DoubleEntryTable__Column--Col {
@@ -84,7 +84,7 @@ export const Styles = styled.div`
       ${TYPOGRAPHY.fontStyles.regular}
       background: ${COLORS.background1};
       border-top-width: 0;
-      border-left-width: ${pxToRem(2)};
+      border-left-width: var(--border-width);
 
       position: sticky;
       left: 0;
@@ -107,8 +107,8 @@ export const Styles = styled.div`
 
       &:first-child {
         background: ${COLORS.background2};
-        border-color: ${COLORS.line1};
-        border-left-width: ${pxToRem(2)};
+        border-color: var(--color-grey-400);
+        border-left-width: var(--border-width);
 
         position: sticky;
         left: 0;