+++ CHANGELOG.md
@@ -5,6 +5,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Feature: Start using Design Tokens for colors config (with fallback).
+- Fix: `ContributionCard`: Remove buggy `borderWidth` prop.
+- Fix: `ToggleSwith`: Modernize components styles.
+- Feature: Design: Used `--border-width` and used A11Y color standards.
+- Feature: Token: Add `--border-width` variable.
+- Feature: Tooling: Add Figma Tokens automatic integration.
+- Feature: `DashboardMenu`: Add `size` prop to `DashboardMenu.Expandable`, defaults to `default`.
+- Fix: `DashboardMenu`: Style fixes.
 
 ## [6.1.0] - 2021-11-15